% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_bank.R
\name{add_phrase_info2}
\alias{add_phrase_info2}
\title{Segment a note track by adding phraseend and phrasebeg columns with boolean markers.}
\usage{
add_phrase_info2(note_track)
}
\arguments{
\item{note_track}{a data frame with an "onset" column}
}
\value{

}
\description{
Segment a note track by adding phraseend and phrasebeg columns with boolean markers.
}
