% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz.R
\docType{data}
\name{Berkowitz}
\alias{Berkowitz}
\title{The Berkowitz sight-singing corpus}
\format{
A data frame with 92789 rows and 22 variables:
\describe{
  \item{melody}{melody in relative (interval) format}
}
}
\source{
\url{ISBN: 978-0-393-28491-1}
}
\usage{
Berkowitz
}
\description{
A dataset containing melodies from the Berkowitz sight-singing corpus
}
\keyword{datasets}
