% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz_files_db.R
\docType{data}
\name{Berkowitz_files_db}
\alias{Berkowitz_files_db}
\title{The Berkowitz sight-singing corpus}
\format{
A data frame with 630 rows and 5 variables:
\describe{
  \item{melody}{melody in relative (interval) format}
}
}
\source{
\url{ISBN: 978-0-393-28491-1}
}
\usage{
Berkowitz_files_db
}
\description{
A dataset containing melodies from the Berkowitz sight-singing corpus
and references to file paths for rhythmic playback
}
\keyword{datasets}
