% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cents}
\alias{cents}
\title{Get cents between two notes}
\usage{
cents(notea, noteb)
}
\arguments{
\item{notea}{The frequency in Hz of note a.}

\item{noteb}{The frequency in Hz of note b.}
}
\value{

}
\description{
Get cents between two notes
}
