% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANTASTIC_Feature_Value_Summary_Statistics.R
\name{compute.entropy}
\alias{compute.entropy}
\title{Compute entropy}
\usage{
\method{compute}{entropy}(integer.vector, alphabet.size = 2)
}
\arguments{
\item{alphabet.size}{}
}
\value{

}
\description{
Compute entropy
}
