% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_audio_feature_item_bank.R
\name{create_audio_feature_bank}
\alias{create_audio_feature_bank}
\title{Build audio features from a folder (wav/mp3)}
\usage{
create_audio_feature_bank(
  audio_file_dir,
  slice_head = NULL,
  reencode_audio_files_to_lame_mp3 = FALSE,
  verbose = FALSE,
  normalise_to_wav = TRUE,
  audio_normalisation_pars = list(target_lufs = -16, lra = 8, tp = -1, trim_silence =
    TRUE, sr = NULL, mono = FALSE)
)
}
\description{
Build audio features from a folder (wav/mp3)
}
