% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_item_bank_into_ngrams.R
\name{create_ngram_item_bank}
\alias{create_ngram_item_bank}
\title{Create ngram item bank}
\usage{
create_ngram_item_bank(
  df,
  lower_ngram_bound = 3L,
  upper_ngram_bound = NULL,
  get_ngrukkon = TRUE,
  get_features = TRUE
)
}
\arguments{
\item{get_features}{}
}
\description{
Create ngram item bank
}
