% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prefix.R
\name{deprefix_cols}
\alias{deprefix_cols}
\title{Removes leading META_/MEL_/AUD_ once (no recursion).}
\usage{
deprefix_cols(df)
}
\arguments{
\item{df}{}
}
\description{
Removes leading META_/MEL_/AUD_ once (no recursion).
}
