% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_audio_feature_item_bank.R
\name{extract_audio_features}
\alias{extract_audio_features}
\title{Extract audio features from an audio file}
\usage{
extract_audio_features(
  audio_file_path,
  reencode_audio_file_to_lame_mp3 = FALSE,
  reencode_audio_file_to_lame_mp3_verbose = FALSE
)
}
\arguments{
\item{reencode_audio_file_to_lame_mp3_verbose}{}
}
\description{
Extract audio features from an audio file
}
