% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_melody_features.R
\name{get_interval_entropy}
\alias{get_interval_entropy}
\title{Get interval entropy}
\usage{
get_interval_entropy(rel_melody_col, sep = ",")
}
\arguments{
\item{sep}{}
}
\value{

}
\description{
Get interval entropy
}
