% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_last_char_of_string}
\alias{get_last_char_of_string}
\title{Grab the last character of a string}
\usage{
get_last_char_of_string(string)
}
\arguments{
\item{string}{}
}
\value{

}
\description{
Grab the last character of a string
}
