% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{get_melodic_feature_table}
\alias{get_melodic_feature_table}
\title{Produce a table describing and operationalising melodic features (for paper-writing)}
\usage{
get_melodic_feature_table(features = NULL)
}
\arguments{
\item{features}{}
}
\value{

}
\description{
Produce a table describing and operationalising melodic features (for paper-writing)
}
