% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_melody_features.R
\name{get_melody_features}
\alias{get_melody_features}
\title{For a db of melodies, get melodic features}
\usage{
get_melody_features(
  df,
  mel_sep = ",",
  durationMeasures = TRUE,
  abs_melody_name = "abs_melody",
  print_histogram = FALSE
)
}
\arguments{
\item{df}{with column "melody" as a relative melody e.g, 2, 2, -1, 3 and "freq" a count of the number of occurrences of the dataset from which it came}

\item{mel_sep}{What is the melody note separator?}

\item{durationMeasures}{Should duration measures be computed?}

\item{print_histogram}{Print histogram of results?}
}
\description{
For a db of melodies, get melodic features
}
