% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_item_bank_into_ngrams.R
\name{get_ngrams_multiple_sizes}
\alias{get_ngrams_multiple_sizes}
\title{Get ngrams of multiple sizes}
\usage{
get_ngrams_multiple_sizes(abs_melody, M)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
Get ngrams of multiple sizes
}
