% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hist_item_bank}
\alias{hist_item_bank}
\title{Plot histograms of the item bank vars/features}
\usage{
hist_item_bank(item_bank, nrow = NULL, ncol = NULL)
}
\arguments{
\item{ncol}{}
}
\value{

}
\description{
Plot histograms of the item bank vars/features
}
