% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_bank.R
\name{midi_file_to_notes_and_durations}
\alias{midi_file_to_notes_and_durations}
\title{Take a MIDI file and return dataframe of the corresponding notes and durations}
\usage{
midi_file_to_notes_and_durations(
  midi_file,
  relativeMelodies = TRUE,
  relativeDurations = FALSE,
  ticks = 480,
  prefix = NULL
)
}
\arguments{
\item{prefix}{}
}
\value{

}
\description{
Take a MIDI file and return dataframe of the corresponding notes and durations
}
