% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midi_to_pitch_class}
\alias{midi_to_pitch_class}
\title{Convert MIDI note to pitch class}
\usage{
midi_to_pitch_class(midi_note)
}
\arguments{
\item{midi_note}{}
}
\value{

}
\description{
Convert MIDI note to pitch class
}
