% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midi_to_pitch_class_numeric}
\alias{midi_to_pitch_class_numeric}
\title{Convert MIDI note to pitch class, in numeric form}
\usage{
midi_to_pitch_class_numeric(midi_note)
}
\arguments{
\item{midi_note}{}
}
\value{

}
\description{
Convert MIDI note to pitch class, in numeric form
}
