% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midi_to_sci_notation}
\alias{midi_to_sci_notation}
\title{Convert MIDI note to its corresponding representation as musical scientific notation}
\usage{
midi_to_sci_notation(midi_note)
}
\arguments{
\item{midi_note}{}
}
\value{

}
\description{
Convert MIDI note to its corresponding representation as musical scientific notation
}
