% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pitch_class_to_midi_notes}
\alias{pitch_class_to_midi_notes}
\title{Return MIDI note(s) corresponding to a given pitch class}
\usage{
pitch_class_to_midi_notes(pitch_class)
}
\arguments{
\item{pitch_class}{}
}
\description{
Return MIDI note(s) corresponding to a given pitch class
}
