% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pitch_class_to_numeric_pitch_class}
\alias{pitch_class_to_numeric_pitch_class}
\title{Convert pitch classes to numeric pitch classes}
\usage{
pitch_class_to_numeric_pitch_class(pitch_class)
}
\arguments{
\item{pitch_class}{}
}
\description{
Convert pitch classes to numeric pitch classes
}
