% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_item_banks.R
\name{plot_categoricals}
\alias{plot_categoricals}
\title{Plot categorical proportions by item bank}
\usage{
plot_categoricals(proportions_tbl)
}
\arguments{
\item{proportions_tbl}{Tibble of feature × bank × level proportions (e.g., res$categoricals$proportions)}
}
\description{
Plot categorical proportions by item bank
}
