% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_item_banks.R
\name{plot_joint_pca_density}
\alias{plot_joint_pca_density}
\title{Plot joint PCA score densities by item bank}
\usage{
plot_joint_pca_density(score_summary)
}
\arguments{
\item{score_data}{Tibble of PCA component scores or summary (e.g., res$pca$joint_score_summary)}
}
\description{
Plot joint PCA score densities by item bank
}
