% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_item_banks.R
\name{plot_violin_summary}
\alias{plot_violin_summary}
\title{Plot feature summaries by item bank and family}
\usage{
plot_violin_summary(summary_data, max_features = 36)
}
\description{
Plot feature summaries by item bank and family
}
