% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{produce_arrhythmic_durations}
\alias{produce_arrhythmic_durations}
\title{Produce arrhythmic durations for a given melody}
\usage{
produce_arrhythmic_durations(x, dur = 0.5)
}
\arguments{
\item{dur}{}
}
\description{
Produce arrhythmic durations for a given melody
}
