% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rel_to_abs_mel}
\alias{rel_to_abs_mel}
\title{Convert a relative melody to an absolute melody}
\usage{
rel_to_abs_mel(rel_mel, start_note = 60)
}
\arguments{
\item{start_note}{}
}
\description{
Convert a relative melody to an absolute melody
}
