% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_last_char_of_string}
\alias{remove_last_char_of_string}
\title{Remove the last character of a string}
\usage{
remove_last_char_of_string(string)
}
\arguments{
\item{string}{}
}
\value{

}
\description{
Remove the last character of a string
}
