% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_bank.R
\name{remove_redundancy}
\alias{remove_redundancy}
\title{Remove redundancy from an item bank}
\usage{
remove_redundancy(
  remove_redundancy,
  item_bank,
  distinct_based_on_melody_only = TRUE
)
}
\arguments{
\item{distinct_based_on_melody_only}{}
}
\description{
Remove redundancy from an item bank
}
