% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_bank.R
\name{save_item_bank}
\alias{save_item_bank}
\title{Save an item bank}
\usage{
save_item_bank(
  save_item_bank_to_file,
  item_bank,
  name,
  type = c("item", "phrase", "ngram", "combined")
)
}
\arguments{
\item{type}{}
}
\description{
Save an item bank
}
