% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_durations_to_have_min_abs_value_of_x_seconds}
\alias{scale_durations_to_have_min_abs_value_of_x_seconds}
\title{Make sure melodies have a minimum absolute value of x (and scale the rest of the durations to this minimum value).}
\usage{
scale_durations_to_have_min_abs_value_of_x_seconds(df, x = 0.25)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Make sure melodies have a minimum absolute value of x (and scale the rest of the durations to this minimum value).
}
