% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sci_notation_to_midi}
\alias{sci_notation_to_midi}
\title{Convert musical scientific notation note to its corresponding representation as a MIDI note}
\usage{
sci_notation_to_midi(sci_notation)
}
\arguments{
\item{sci_notation}{}
}
\description{
Convert musical scientific notation note to its corresponding representation as a MIDI note
}
