% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_item_bank_in_db.R
\name{store_item_bank_in_db}
\alias{store_item_bank_in_db}
\title{Store an item bank in the musicassessr database}
\usage{
store_item_bank_in_db(
  item_bank,
  item_bank_name,
  item_bank_description,
  db_name = Sys.getenv("DB_NAME"),
  db_host = Sys.getenv("DB_HOST"),
  db_username = Sys.getenv("DB_USER"),
  db_password = Sys.getenv("DB_PASSWORD")
)
}
\arguments{
\item{db_password}{}
}
\description{
Store an item bank in the musicassessr database
}
