% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_mel_to_vector}
\alias{str_mel_to_vector}
\title{Convert a string of a melody, separated by a to a vector}
\usage{
str_mel_to_vector(str_mel, sep = ",", numeric = TRUE)
}
\arguments{
\item{numeric}{}
}
\description{
Convert a string of a melody, separated by a to a vector
}
