% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unclass_item_bank}
\alias{unclass_item_bank}
\title{Unclass item bank class (to tibble)}
\usage{
unclass_item_bank(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Unclass item bank class (to tibble)
}
