% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_cents}
\alias{vector_cents}
\title{Get cents between vector and reference note}
\usage{
vector_cents(reference_note, vector_of_values)
}
\arguments{
\item{vector_of_values}{}
}
\value{

}
\description{
Get cents between vector and reference note
}
