% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_cents_between_two_vectors}
\alias{vector_cents_between_two_vectors}
\title{Get the cents between two vectors}
\usage{
vector_cents_between_two_vectors(vectora, vectorb)
}
\arguments{
\item{vectorb}{}
}
\value{

}
\description{
Get the cents between two vectors
}
