% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_bank.R
\name{corpus_to_item_bank}
\alias{corpus_to_item_bank}
\title{Convert a corpus to a useful item bank}
\usage{
corpus_to_item_bank(
  corpus_name,
  midi_file_dir = NULL,
  musicxml_file_dir = NULL,
  prefix = NULL,
  corpus_df = NULL,
  phrases_db = NULL,
  output_type = c("both", "ngram", "phrases"),
  launch_app = TRUE,
  main_db_combine_ngram_and_phrases = TRUE
)
}
\arguments{
\item{main_db_combine_ngram_and_phrases}{}
}
\value{

}
\description{
Convert a corpus to a useful item bank
}
