% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{get_melodic_feature_table}
\alias{get_melodic_feature_table}
\title{Produce a table describing and operationalising melodic features (for paper-writing)}
\usage{
get_melodic_feature_table(features = NULL, return_kable = TRUE)
}
\arguments{
\item{features}{Optional character vector of feature names to filter.}

\item{return_kable}{If TRUE (default), return knitr::kable with escape = FALSE.}
}
\value{
tibble or knitr_kable
}
\description{
Produce a table describing and operationalising melodic features (for paper-writing)
}
\examples{
get_melodic_feature_table(c("target_melody_length","i.entropy","tonalness"))
}
