% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_melody_features.R
\name{get_melody_features}
\alias{get_melody_features}
\title{For a db of melodies, get melodic features}
\usage{
get_melody_features(
  df,
  mel_sep = ",",
  durationMeasures = TRUE,
  abs_melody_name = "abs_melody"
)
}
\arguments{
\item{df}{with column "melody" as a relative melody e.g, 2, 2, -1, 3 and "freq" a count of the number of occurrences of the dataset from which it came}

\item{durationMeasures}{}
}
\value{

}
\description{
For a db of melodies, get melodic features
}
