% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_item_bank.R
\name{midi_file_to_notes_and_durations}
\alias{midi_file_to_notes_and_durations}
\title{Get notes and durations from a MIDI file}
\usage{
midi_file_to_notes_and_durations(midi_file, prefix = NULL)
}
\arguments{
\item{prefix}{}
}
\value{

}
\description{
Get notes and durations from a MIDI file
}
