% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phrases_item_bank.R
\name{segment_phrase}
\alias{segment_phrase}
\title{Segment a note track by adding phrasend and phrasbeg columns with boolean markers.}
\usage{
segment_phrase(
  note_track,
  phrase_segment_outlier_threshold = 0.65,
  ioi_threshold = 0.96
)
}
\arguments{
\item{note_track}{a data frame with an "onset" column}

\item{phrase_segment_outlier_threshold}{A threshold for selecting outliers for durations in phrase segmentation.}

\item{ioi_threshold}{A threshold for selecting outliers for iois in phrase segmentation.}
}
\description{
Segment a note track by adding phrasend and phrasbeg columns with boolean markers.
}
