% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_item_bank.R
\name{subset_item_bank}
\alias{subset_item_bank}
\title{Subset Item Bank}
\usage{
subset_item_bank(
  item_bank,
  item_length = NULL,
  quantile_cut = if ("log_freq" \%in\% names(item_bank)) min(item_bank$log_freq) else
    -Inf,
  span_min = min(item_bank$span),
  span_max = max(item_bank$span),
  tonality = NULL,
  min_mean_duration = 0,
  return_as_item_bank_class = FALSE,
  retain_classes = FALSE
)
}
\arguments{
\item{retain_classes}{Should the item bank keep any extra classes it had?}
}
\description{
Subset Item Bank
}
