% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_item_bank.R
\name{subset_item_bank}
\alias{subset_item_bank}
\title{Subset Item Bank}
\usage{
subset_item_bank(
  item_bank,
  item_length = NULL,
  quantile_cut = -Inf,
  span_min = 0L,
  span_max = Inf,
  tonality = NULL,
  min_mean_duration = 0,
  return_as_item_bank_class = FALSE,
  retain_classes = FALSE
)
}
\arguments{
\item{retain_classes}{Should the item bank keep any extra classes it had?}
}
\description{
Subset Item Bank
}
